Withdraw the dtype modification.

+++ tests/python/deepep/test_common.py
@@ -79,8 +79,7 @@ def normal_test(
             x
             * topk_weights_recv.masked_fill(topk_idx == -1, 0)
             .sum(dim=1)
-            .view(-1, 1)
-            .to(x.dtype),
+            .view(-1, 1),
         )
         < 5e-5
     )
